reorganized files into app and test folder
- main.py models and helper functions in app folder
- test files in test folder
